Migrate

+++ database/migrations/2024_12_17_120452_create_comments_table.php
@@ -15,12 +15,12 @@ return new class extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->increments('id'); // Auto-incrementing primary key
-            $table->unsignedBigInteger('recipe_id'); // Integer to reference the recipe
-            $table->unsignedBigInteger('user_id'); // Integer to reference the user
+            $table->unsignedInteger('recipe_id'); // Foreign key to recipes
+            $table->unsignedBigInteger('user_id'); // Foreign key to users
             $table->text('comment'); // TEXT for the comment
             $table->timestamps(); // Adds "created_at" and "updated_at" columns
 
-            // Optional: Add foreign key constraints
+            // Foreign key constraints
             $table->foreign('recipe_id')->references('id')->on('recipes')->onDelete('cascade');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
